requirements.txt
django
mysqlclient
djoser
djangorestframework-simplejwt
pillow
django-cors-headers
boto3 
django-storages